fix(db): Flyway V1과 공유 모델 정합성을 맞춤
- works의 제거된 status 매핑과 BIGINT 타입 불일치 수정
- 공유 모델의 생성 및 수정 시각을 NOT NULL 자동 기록으로 통일
- 청크와 설정 후보 Mapper의 nullable timestamp 처리 제거
- Flyway schema 계약 테스트 추가

+++ app/models/upload_file.py
@@ -1,6 +1,6 @@
 from uuid import UUID
 
-from sqlalchemy import Integer, String
+from sqlalchemy import BigInteger, Integer, String
 from sqlalchemy.dialects.postgresql import UUID as PostgresUUID
 from sqlalchemy.orm import Mapped, mapped_column
 
@@ -17,7 +17,7 @@ class UploadFile(TimestampMixin, Base):
     original_filename: Mapped[str] = mapped_column(String(255))
     mime_type: Mapped[str | None] = mapped_column(String(100))
     storage_url: Mapped[str | None] = mapped_column(String(512))
-    file_size: Mapped[int] = mapped_column(Integer)
+    file_size: Mapped[int] = mapped_column(BigInteger)
     detected_episode_start_no: Mapped[int | None] = mapped_column(Integer)
     detected_episode_end_no: Mapped[int | None] = mapped_column(Integer)
     detected_episode_count: Mapped[int | None] = mapped_column(Integer)
